fix(bet_history): leaderboard number type double

+++ lib/bet_history/screens/bet_history_page.dart
@@ -122,7 +122,7 @@ class BetHistoryRow extends StatelessWidget {
 String totalRisk({List<int> firstList, List<int> secondList}) {
   final firstListSum = firstList.isEmpty
       ? 0
-      : firstList.reduce((value, element) => value + element).toDouble();
+      : firstList.reduce((value, element) => value + element);
   // final secondListSum = secondList.isEmpty
   //     ? 0
   //     : secondList.reduce((value, element) => value + element).toDouble();
